Refactor code to move repeated code into a file
